added observable pattern for content notifications

+++ P1/src/main/java/com/example/P1/service/UserService.java
@@ -3,6 +3,8 @@ package com.example.P1.service;
 import com.example.P1.model.BillingDetails;
 import com.example.P1.model.ItemNotFoundException;
 import com.example.P1.model.User;
+import com.example.P1.notifications.Observable;
+import com.example.P1.notifications.Observer;
 import com.example.P1.repository.UserConnectionDB;
 import com.example.P1.contract.UserConnectionContract;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -22,10 +24,12 @@ public class UserService implements UserConnectionContract {
      * and used for interface dependency injection
      */
     private final UserConnectionDB userConnectionDB;
+    private final Observable contentService;
 
     @Autowired
-    public UserService(UserConnectionDB userConnectionDB) {
+    public UserService(UserConnectionDB userConnectionDB, Observable contentService) {
         this.userConnectionDB = userConnectionDB;
+        this.contentService = contentService;
     }
 
     @Override
@@ -41,12 +45,17 @@ public class UserService implements UserConnectionContract {
     @Override
     public User addUser(User user) {
         user.setId(UUID.randomUUID().toString());
+        contentService.registerObserver((Observer) user);
         return userConnectionDB.save(user);
     }
 
     @Override
     public void deleteUserById(String id) {
-        userConnectionDB.deleteById(id);
+        User  user = userConnectionDB.findUserById(id).orElse(null);
+        if(user != null){
+            contentService.removeObserver(user);
+            userConnectionDB.deleteById(id);
+        }
     }
 
     @Override
@@ -58,4 +67,6 @@ public class UserService implements UserConnectionContract {
     public BillingDetails getLastPaymentOfUser(String id) {
         return userConnectionDB.getLastPaymentOfUser(id).orElseThrow(() -> new ItemNotFoundException("User by id " + id + " has no recorded payments"));
     }
+
+    //notifications methods
 }
